pyproject.toml
[project]
name = "ch-blog-sync"
version = "0.1.0"
description = "Automated system to sync ClickHouse blogs from RSS feed"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.11",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "pyyaml>=6.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.uv]
dev-dependencies = []
